requirements.txt
Flask==1.1.1
Flask-JWT==0.3.2
Flask-RESTful==0.3.7
psycopg2-binary==2.8.3
openpyxl==3.0.0
requests==2.18.4
Werkzeug==0.16.0